Update targetOutputBase.py
Updated the module so that it imports the moleculeName variable from the matlabProcessing module so that it is not necessary to always update the molecule name. Also, changed the code on line 61 so that the reshape parameter for the number of columns does not have to be hard-coded, replaced by len(originalMatrix[0]).