Update moleculePlotting.py
Completely updated the module so that the two molecules (original molecule in blue, solution molecule in red) are shown as line structures rather than just dots. This makes it easier to visualize and compare structures. It is important to note that for essentially linear structures (e.g. C2H2) the deviations are so small that the axis graduations will be very small, sometimes on the order of 1e-10. Even though the solution molecule may look radically deviated, the changes in structure are not as dramatic as they seem.